backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv
google-genai
google-api-core
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-chroma
chromadb
unstructured[md]
tenacity
numpy<2.0.0
